% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_pie.R
\name{cgm_pie}
\alias{cgm_pie}
\title{Pie Chart}
\usage{
cgm_pie(cgm.data, file.name = NULL)
}
\arguments{
\item{cgm.data}{A dataframe of CGM data formatted by format_dexcom() for a patient, containing all timepoints}

\item{file.name}{The name of the output file}
}
\description{
Generate pie chart for within, below, and above range percentage
}
