% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfa.functions.R
\name{convert.to.random.walk}
\alias{convert.to.random.walk}
\title{DFA Random Walk (Step 1)}
\usage{
convert.to.random.walk(glucose)
}
\arguments{
\item{glucose}{A vector of glucose values from CGM data}
}
\description{
Convert a vector into a random walk
}
