% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiseCGM.R
\docType{data}
\name{grobman_Pt2_T2}
\alias{grobman_Pt2_T2}
\title{Continuous Glucose Monitoring Data from October}
\format{A list of 4 data frames.}
\usage{
data(grobman_Pt2_T2)
}
\description{
A list of datasets containing the data from a DEXCOM CGM Device and the meta information that is returned from the format_dexcom() function:
}
\details{
\itemize{
  \item full.cgm.data. A dataframe of 2947 rows and 20 columns
  \item cleaned.cgm.data. A dataframe of 2845 rows and 22 columns. This does not contain data for any days if the device needed calibration but was not performed. Missing data are imputed with mean of 2 measurements before and after the missing data.
  \item meta.data. A dataframe containing the meta information provided by DEXCOM devices
  \item informative.meta.data. A dataframe containing the High, Low, and Urgent Low threshold.
}
}
\keyword{datasets}
