% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_boxplot.R
\name{hourly_boxplot}
\alias{hourly_boxplot}
\title{Hourly Boxplot}
\usage{
hourly_boxplot(cgm.data, informative.cgm.meta, file.name = NULL)
}
\arguments{
\item{cgm.data:}{A dataframe of CGM data formatted by format_dexcom() for a patient, containing all timepoints}

\item{informative.cgm.meta:}{A dataframe of CGM meta data formatted by format_dexcom()}

\item{file.name:}{The name of the output file}
}
\description{
Generate hourly boxplot
}
