% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfa.functions.R
\name{local.fluctuation.hurst}
\alias{local.fluctuation.hurst}
\title{Local Fluctuation (Step 4)}
\usage{
local.fluctuation.hurst(cgm.rw, multi.dfa.test, scale_small)
}
\arguments{
\item{cgm.rw}{A dataframe of random walk values}

\item{multi.dfa.test}{A multifractual DFA analysis output}

\item{scale_small}{A vector of "small" scales to determine bin sizes for each analyses}
}
\description{
Performed monofractual DFA for a givien dataframe of random walk values and a vector of scales
}
