% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfa.functions.R
\name{mono.dfa}
\alias{mono.dfa}
\title{Monofractual Detrended Fluctuation Analysis (Step 2)}
\usage{
mono.dfa(cgm.rw, scale.vect)
}
\arguments{
\item{cgm.rw}{A dataframe of random walk values}

\item{scale.vect}{A vector of scales to determine bin sizes for each analyses}
}
\description{
Performed monofractual DFA for a givien dataframe of random walk values and a vector of scales
}
