% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfa.functions.R
\name{multi.dfa}
\alias{multi.dfa}
\title{Multifractual Detrended Fluctuation Analysis (Step 3)}
\usage{
multi.dfa(mono.dfa.test, q.vect, scale.vect)
}
\arguments{
\item{mono.dfa.test}{A monofractual DFA analysis output}

\item{q.vect}{A vector of scales for the q-th order statistics}

\item{scale.vect}{A vector of scales to determine bin sizes for each analyses}
}
\description{
Performed monofractual DFA for a givien dataframe of random walk values and a vector of scales
}
