% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_time_series_plot.R
\name{stacked_time_series_plot}
\alias{stacked_time_series_plot}
\title{Stacked Time Series Plot}
\usage{
stacked_time_series_plot(cgm.data, informative.cgm.meta,
  file.name = NULL)
}
\arguments{
\item{cgm.data:}{A dataframe of CGM data formatted by format_dexcom() for a patient, containing all timepoints}

\item{informative.cgm.meta:}{A dataframe of CGM meta data formatted by format_dexcom()}

\item{file.name:}{The name of the output file}
}
\description{
Generate Time Series Plot from all three time points stacked on top of each other
}
